Get rid module requires
